chore: rename `Linter` -> `LinterPlugin`
