+++ src/cloud-element-templates/LinterPlugin.js
